+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.x branch
 ## 2.23 branch
 ### 2.23.2
+* `AppValidator` moved from `MeCms\Model\Validation` to `MeCms\Validation`;
 * updated for CakePHP 3.6;
 * updated for cakephp-database-backup 2.5.0.